// src/app/database/interfaces/idb-table.model.ts
import Dexie, { Collection } from 'dexie';
import { EntityStateEnum } from '../entity-state.enum';

export interface ITableSchema {
  name: string;
  schema: string;
}

export interface IDexieTableSchema {
  name: string;
  primKey: { src: string };
  indexes: { src: string }[];
}

export interface IFilterDelegate {
  (dbSet: Dexie.Table): Collection;
}

export interface IEntitySyncDTO {
  Entity: object;
  State: EntityStateEnum;
  Table: string;
}